feat: share online user count via context
